Semi-working fasing 24bit and USI state machine
Using lookup tables to generate colour. Green works. Indigo might
be the right colour but dimms a bit toward the brighter end. All
other colours look like a working version of Indigo.

USI state machine is implimented and I beleive working!

The look of the USI vs the pattern code is more final now, except
the pattern code isn't in a header file yet, it's marked with a
TODO.