Improve colours
Having only one 0 in each table seemed sensible but it's wrong. I
did have to move the 0 to 1 boundaries around to sort of line up
with the 255 table though to prevent flashed of primary colour
before the LED turned off.

Alse got the perl script to write all the code.